requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
alabaster==0.7.13
    # via sphinx
anyascii==0.3.2
    # via wagtail
asgiref==3.6.0
    # via django
babel==2.12.1
    # via sphinx
beautifulsoup4==4.11.2
    # via wagtail
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
colorama==0.4.6
    # via sphinx-autobuild
crispy-bootstrap4==2022.1
    # via -r requirements.in
dj-database-url==1.2.0
    # via environs
dj-email-url==1.0.6
    # via environs
django==4.1.7
    # via
    #   -r requirements.in
    #   crispy-bootstrap4
    #   dj-database-url
    #   django-appconf
    #   django-auth-ldap
    #   django-crispy-forms
    #   django-csp
    #   django-csp-reports
    #   django-extensions
    #   django-filter
    #   django-modelcluster
    #   django-permissionedforms
    #   django-ranged-response
    #   django-select2
    #   django-simple-captcha
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   wagtail
django-appconf==1.0.5
    # via django-select2
django-auth-ldap==4.1.0
    # via -r requirements.in
django-cache-url==3.4.4
    # via environs
django-crispy-forms==2.0
    # via
    #   -r requirements.in
    #   crispy-bootstrap4
django-csp==3.7
    # via -r requirements.in
django-csp-reports==1.8.1
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-filter==22.1
    # via wagtail
django-ipware==5.0.0
    # via -r requirements.in
django-modelcluster==6.0
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-private-storage==3.0
    # via -r requirements.in
django-ranged-response==0.2.0
    # via django-simple-captcha
django-select2==8.1.1
    # via -r requirements.in
django-simple-captcha==0.5.17
    # via -r requirements.in
django-sphinx-view==22.1a6
    # via -r requirements.in
django-taggit==3.1.0
    # via wagtail
django-treebeard==4.6.1
    # via wagtail
djangorestframework==3.14.0
    # via wagtail
docutils==0.19
    # via
    #   myst-parser
    #   sphinx
draftjs-exporter==2.1.7
    # via wagtail
environs[django]==9.5.0
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
html5lib==1.1
    # via wagtail
huey==2.4.5
    # via -r requirements.in
idna==3.4
    # via requests
imagesize==1.4.1
    # via sphinx
jinja2==3.1.2
    # via
    #   myst-parser
    #   sphinx
l18n==2021.3
    # via wagtail
livereload==2.6.3
    # via sphinx-autobuild
markdown-it-py==2.2.0
    # via
    #   mdit-py-plugins
    #   myst-parser
markupsafe==2.1.2
    # via jinja2
marshmallow==3.19.0
    # via environs
mdit-py-plugins==0.3.5
    # via myst-parser
mdurl==0.1.2
    # via markdown-it-py
myst-parser==1.0.0
    # via -r requirements.in
openpyxl==3.1.2
    # via wagtail
packaging==23.0
    # via
    #   marshmallow
    #   sphinx
pillow==9.4.0
    # via
    #   django-simple-captcha
    #   wagtail
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   python-ldap
pyasn1-modules==0.2.8
    # via python-ldap
pyclamd==0.4.0
    # via -r requirements.in
pygments==2.14.0
    # via sphinx
python-dotenv==1.0.0
    # via environs
python-ldap==3.4.3
    # via django-auth-ldap
python-magic==0.4.27
    # via -r requirements.in
pytz==2022.7.1
    # via
    #   django-modelcluster
    #   djangorestframework
    #   l18n
pyyaml==6.0
    # via myst-parser
requests==2.28.2
    # via
    #   sphinx
    #   wagtail
six==1.16.0
    # via
    #   html5lib
    #   l18n
    #   livereload
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.4
    # via beautifulsoup4
sphinx==6.1.3
    # via
    #   -r requirements.in
    #   myst-parser
    #   sphinx-autobuild
sphinx-autobuild==2021.3.14
    # via -r requirements.in
sphinxcontrib-applehelp==1.0.4
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.1
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlparse==0.4.3
    # via django
telepath==0.3
    # via wagtail
tornado==6.2
    # via livereload
urllib3==1.26.15
    # via requests
wagtail==4.2.1
    # via -r requirements.in
webencodings==0.5.1
    # via html5lib
willow==1.4.1
    # via wagtail
